Added new components and improved the code.
