+++ client/src/components/AttractionsGrid.jsx
@@ -58,19 +58,21 @@ const AttractionCard = ({ title, description, image }) => {
 
 const AttractionsGrid = () => {
   return (
-    <div className="container mx-auto px-4 py-8 max-w-screen-xl overflow-hidden">
-      <h2 className="text-4xl font-bold text-center mb-8 text-gray-800">
-        Discover Andhra Pradesh
-      </h2>
-      <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6 w-full">
-        {attractionsData.map((attraction, index) => (
-          <AttractionCard 
-            key={index}
-            title={attraction.title}
-            description={attraction.description}
-            image={attraction.image}
-          />
-        ))}
+    <div className="relative z-10 bg-white">
+      <div className="container mx-auto px-4 py-8 max-w-screen-xl overflow-hidden">
+        <h2 className="text-4xl font-bold text-center mb-8 text-gray-800">
+          Discover Andhra Pradesh
+        </h2>
+        <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6 w-full">
+          {attractionsData.map((attraction, index) => (
+            <AttractionCard 
+              key={index}
+              title={attraction.title}
+              description={attraction.description}
+              image={attraction.image}
+            />
+          ))}
+        </div>
       </div>
     </div>
   );
