+++ client/src/main.jsx
@@ -1,25 +1,14 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import { BrowserRouter, Routes, Route, Outlet } from "react-router";
+import { BrowserRouter, Routes, Route } from "react-router";
 import App from "./App";
 import "./index.css";
 import AIChatBot from "./pages/Bot";
-import Header from "./components/Header";
-import Footer from "./components/Footer";
 import TourismMetricsDashboard from "./pages/TourismMetric";
+import Layout from "./components/Layout";
 
 // Layout component with persistent Header
-const Layout = () => {
-  return (
-    <div className="flex flex-col min-h-screen overflow-hidden">
-      <Header />
-      <main className="flex-grow mt-20">
-        <Outlet />
-      </main>
-      <Footer />
-    </div>
-  );
-};
+
